Registro Llegada incompleto

+++ ClinicaFrba/ClinicaFrba/BD/Entidades/Turno.cs
@@ -13,6 +13,6 @@ namespace ClinicaFrba.BD.Entidades
         public string especialidad { get; set; }
         public string nombre_Pac { get; set; }
         public string apellido_Pac { get; set; }
-        public int id_Consulta { get; set; }
+        public decimal id_Consulta { get; set; }
     }
 }
